Make intro more illustrative and beginner-friendly
After reading the old intro paragraph, I had only a vague idea of what CORS is, and I was frustrated by what I saw as a lack of clarity, specificity, and beginner-friendliness in the explanation.  Many hours on WHATWG later, I understand CORS more, and believe I can make this intro better.

I replaced the old intro paragraph with three paragraphs:

The first paragraph describes what CORS is, and gives an example of how CORS works. For the sake of beginner-friendliness, I aimed for specificity, even at the price of concision.

The second paragraph describes what I understand to be the main advantage and the main disadvantage of CORS as it applies to web development. I believe this would be helpful to those trying to understand the purpose and pitfalls of CORS at a glance.

The third paragraph describes how browsers can check servers for CORS compatibility using "CORS-preflight" requests.